build(shop-mvc-express): replace build.js with esbuild.config.mjs and refine config
